Re-implement migration from no_NO to nb_NO
The migration steps in the original implementation of the change
from no_NO to nb_NO are incomplete leaving the database in an
inconsistent state if no_NO was enabled before the upgrade from 2.x to 3.x.

This change will convert the migration to use the infrastructure
originally developed in pkp/pkp-lib#3416 to handle the
sr_SR->sr_RS@latin change.